Re-organise ledger types.
