+++ specs/ledger/hs/src/core/core/Ledger/Core.hs
@@ -1,0 +1,40 @@
+module Ledger.Core where
+
+import Ledger.Signatures
+import Numeric.Natural (Natural)
+
+-- | Hash part of the ledger paylod
+class HasHash a where
+  hash :: a -> Hash
+
+---------------------------------------------------------------------------------
+-- Signing and verification
+---------------------------------------------------------------------------------
+
+-- |Representation of the owner of key pair.
+newtype Owner = Owner Natural deriving (Show, Eq, Ord)
+
+-- |Signing Key.
+newtype SKey = SKey Owner deriving (Show, Eq, Ord)
+
+-- |Verification Key.
+newtype VKey = VKey Owner deriving (Show, Eq, Ord)
+
+-- |Key Pair.
+data KeyPair = KeyPair
+  {sKey :: SKey, vKey :: VKey} deriving (Show, Eq, Ord)
+
+-- |Return a key pair for a given owner.
+keyPair :: Owner -> KeyPair
+keyPair owner = KeyPair (SKey owner) (VKey owner)
+
+-- |A digital signature.
+data Sig a = Sig a Owner deriving (Show, Eq, Ord)
+
+-- |Produce a digital signature
+sign :: SKey -> a -> Sig a
+sign (SKey k) d = Sig d k
+
+-- |Verify a digital signature
+verify :: Eq a => VKey -> a -> Sig a -> Bool
+verify (VKey vk) vd (Sig sd sk) = vk == sk && vd == sd
